% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_504_get.R
\name{error_504_get}
\alias{error_504_get}
\title{error 504 get}
\usage{
error_504_get(url, key)
}
\arguments{
\item{url}{a website url}

\item{key}{an API token}
}
\value{
a GET response
}
\description{
This function helps with timeout errors from the
API. Will rerun the request that has just timed out
after 10 seconds
}
\examples{
#NOT RUN error_504_get("www.timedoutrequest.com", myapikey)
}
