% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_504_post.R
\name{error_504_post}
\alias{error_504_post}
\title{error 504 post}
\usage{
error_504_post(url, key, body)
}
\arguments{
\item{url}{a website url}

\item{key}{an API token}

\item{body}{a body for the POST request}
}
\value{
a POST response
}
\description{
This function helps with timeout errors from the
API. Will rerun the request after 10 seconds
}
\examples{
#NOT RUN error_504_post("www.timedoutrequest.com", body= list(flow=1,variables=list(name=list(format="category"))),myapikey)
}
