% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow.R
\name{get_flow}
\alias{get_flow}
\title{get flow}
\usage{
get_flow(flow = NULL, url, key, sum = TRUE)
}
\arguments{
\item{flow}{a flow uuid}

\item{url}{url formatted: "https:://app.communityconnectlabs.com/api/v2/endpoint.json?flow=" url must end in "flow=" can include groups
and other queries before "flow="}

\item{key}{an API token}

\item{sum}{TRUE/FALSE to sum results or not}
}
\value{
a dataframe
}
\description{
This function runs a GET request for the whole length of a given url and flow.
CCL is rate-limited to 2000 results, so this function will continue to make requests until
the next link is null
}
\examples{
#NOT RUN get_flow(flow=uuid12345, url="https://app.communityconnectlabs.com/api/v2/contacts_report.json?flow=",
#key=myAPIkey)

#NOT RUN - With additional query get_flow(flow=uuid12345,url="https://app.communityconnectlabs.com/api/v2/contacts_report.json?group=Opted-Out&flow=",
#key = myAPIkey, sum = TRUE)

}
