% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_manual.R
\name{get_flow_manual}
\alias{get_flow_manual}
\title{get flow manual}
\usage{
get_flow_manual(url, key, sum = TRUE)
}
\arguments{
\item{url}{url}

\item{key}{an API token}

\item{sum}{TRUE/FALSE if wish to sum values}
}
\value{
a dataframe
}
\description{
Runs an entire flow get request. Pass in a url manually formatted. Particularly useful when
running an endpoint for just 1 flow.
}
\examples{
#NOT RUN
#get_flow_manual("https://app.communityconnectlabs.com/api/v2/contacts_report.json?flow=uuid124",key=myAPIkey)

}
