% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_flow.R
\name{post_flow}
\alias{post_flow}
\title{post flow}
\usage{
post_flow(url, body, key, sum = TRUE)
}
\arguments{
\item{url}{website url with endpoint}

\item{body}{a POST Body}

\item{key}{an API token}

\item{sum}{TRUE/FALSE If results should be summed up}
}
\value{
a Dataframe
}
\description{
This function runs a POST request for the whole length of a given url and flow.
CCL is rate-limited to 2000 results, so this function will continue to make requests until
the next link is null
}
\examples{
#NOT RUN, API token needed  post_flow("https://app.communityconnectlabs.com/api/v2/flow_variable_report.json",
#body = list(flow=5,variables=list(var1 =list(format="category"),var2=list(format="category"))), sum = TRUE, key=myAPIkey)

}
