% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_to_list.R
\name{response_to_list}
\alias{response_to_list}
\title{response_to_list}
\usage{
response_to_list(raw_response)
}
\arguments{
\item{raw_response}{a result of a GET() or POST() call}
}
\value{
a list
}
\description{
Converts a GET or POST request from JSON to a list.
}
\examples{
response_to_list(httr::GET("https://app.communityconnectlabs.com/api/v2/contacts.json"))
}
